AbstractJdbcDestination is implementing the AbstractDestination class
