+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/base/AbstractDestination.java
@@ -1,3 +1,27 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020 Airbyte
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package io.airbyte.integrations.base;
 
 import com.fasterxml.jackson.databind.JsonNode;
@@ -16,6 +40,34 @@ import java.util.Set;
 
 public abstract class AbstractDestination implements Destination, DestinationConsumerCallback {
 
+  /**
+   * Strategy:
+   * <p>
+   * 1. Create a temporary table for each stream
+   * </p>
+   * <p>
+   * 2. Accumulate records in a buffer. One buffer per stream.
+   * </p>
+   * <p>
+   * 3. As records accumulate write them in batch to the database. We set a minimum numbers of records
+   * before writing to avoid wasteful record-wise writes.
+   * </p>
+   * <p>
+   * 4. Once all records have been written to buffer, flush the buffer and write any remaining records
+   * to the database (regardless of how few are left).
+   * </p>
+   * <p>
+   * 5. In a single transaction, delete the target tables if they exist and rename the temp tables to
+   * the final table name.
+   * </p>
+   *
+   * @param config - integration-specific configuration object as json. e.g. { "username": "airbyte",
+   *        "password": "super secure" }
+   * @param catalog - schema of the incoming messages.
+   * @return consumer that writes singer messages to the database.
+   * @throws Exception - anything could happen!
+   */
+  @Override
   public DestinationConsumer<AirbyteMessage> write(JsonNode config, ConfiguredAirbyteCatalog catalog) throws Exception {
     connectDatabase(config);
     final Map<String, WriteConfig> writeBuffers = new HashMap<>();
@@ -44,7 +96,7 @@ public abstract class AbstractDestination implements Destination, DestinationCon
 
   protected abstract void connectDatabase(JsonNode config);
 
-  protected abstract void queryDatabase(String query);
+  protected abstract void queryDatabase(String query) throws Exception;
 
   protected String getDefaultSchemaName(JsonNode config) {
     if (config.has("schema")) {
@@ -61,6 +113,9 @@ public abstract class AbstractDestination implements Destination, DestinationCon
   protected abstract String createRawTableQuery(String schemaName, String streamName);
 
   public abstract void writeQuery(int batchSize, CloseableQueue<byte[]> writeBuffer, String schemaName, String tmpTableName);
-  public abstract void commitRawTables(Map<String, WriteConfig> writeConfigs);
+
+  public abstract void commitRawTables(Map<String, WriteConfig> writeConfigs) throws Exception;
+
   public abstract void cleanupTmpTables(Map<String, WriteConfig> writeConfigs);
+
 }
