--- airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/base/AbstractJdbcIntegration.java
@@ -1,96 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2020 Airbyte
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package io.airbyte.integrations.base;
-
-import static org.jooq.impl.DSL.currentSchema;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import io.airbyte.commons.json.Jsons;
-import io.airbyte.commons.resources.MoreResources;
-import io.airbyte.db.Database;
-import io.airbyte.db.Databases;
-import io.airbyte.protocol.models.AirbyteConnectionStatus;
-import io.airbyte.protocol.models.AirbyteConnectionStatus.Status;
-import io.airbyte.protocol.models.ConnectorSpecification;
-import java.io.IOException;
-import org.jooq.DSLContext;
-import org.jooq.SQLDialect;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public abstract class AbstractJdbcIntegration implements Integration {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(AbstractJdbcIntegration.class);
-
-  private final String driverClass;
-  private final SQLDialect dialect;
-
-  public AbstractJdbcIntegration(final String driverClass, final SQLDialect dialect) {
-    this.driverClass = driverClass;
-    this.dialect = dialect;
-  }
-
-  public abstract JsonNode toJdbcConfig(JsonNode config);
-
-  @Override
-  public ConnectorSpecification spec() throws IOException {
-    // return a JsonSchema representation of the spec for the integration.
-    final String resourceString = MoreResources.readResource("spec.json");
-    return Jsons.deserialize(resourceString, ConnectorSpecification.class);
-  }
-
-  @Override
-  public AirbyteConnectionStatus check(JsonNode config) {
-    try (final Database database = getDatabase(config)) {
-      // attempt to get current schema. this is a cheap query to sanity check that we can connect to the
-      // database. `currentSchema()` is a jooq method that will run the appropriate query based on which
-      // database it is connected to.
-      database.query(this::getCurrentDatabaseName);
-
-      return new AirbyteConnectionStatus().withStatus(Status.SUCCEEDED);
-    } catch (Exception e) {
-      LOGGER.debug("Exception while checking connection: ", e);
-      return new AirbyteConnectionStatus()
-          .withStatus(Status.FAILED)
-          .withMessage("Can't connect with provided configuration.");
-    }
-  }
-
-  protected Database getDatabase(JsonNode config) {
-    final JsonNode jdbcConfig = toJdbcConfig(config);
-
-    return Databases.createDatabase(
-        jdbcConfig.get("username").asText(),
-        jdbcConfig.has("password") ? jdbcConfig.get("password").asText() : null,
-        jdbcConfig.get("jdbc_url").asText(),
-        driverClass,
-        dialect);
-  }
-
-  protected String getCurrentDatabaseName(DSLContext ctx) {
-    return ctx.select(currentSchema()).fetch().get(0).getValue(0, String.class);
-  }
-
-}
