+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/base/WriteConfig.java
@@ -21,6 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
 package io.airbyte.integrations.base;
 
 import io.airbyte.commons.lang.CloseableQueue;
